fix: make sure that we can handle very big numbers

+++ src/main/java/jimlind/filmlinkd/system/letterboxd/model/LBFilmStatisticsCounts.java
@@ -2,10 +2,10 @@ package jimlind.filmlinkd.system.letterboxd.model;
 
 // https://api-docs.letterboxd.com/#/schemas/FilmStatisticsCounts
 public class LBFilmStatisticsCounts {
-  public int watches;
-  public int likes;
-  public int ratings;
-  public int fans;
-  public int lists;
-  public int reviews;
+  public long watches;
+  public long likes;
+  public long ratings;
+  public long fans;
+  public long lists;
+  public long reviews;
 }
